specify dtype for empty dataframes
pandas throws a warning when using concat on two data frames with non-matching dtypes. This is also the case for empty data frames. This commit addresses the issue by defining the dtype.

Addtionally, applymap was replaced with map as applymap is deprecated.

+++ plspm/bootstrap.py
@@ -22,7 +22,7 @@ from plspm.weights import WeightsCalculatorFactory
 from plspm.estimator import Estimator
 
 def _create_summary(data: pd.DataFrame, original):
-    summary = pd.DataFrame(0, index=data.columns, columns=["original", "mean", "std.error", "perc.025", "perc.975", "t stat."])
+    summary = pd.DataFrame(0.0, index=data.columns, columns=["original", "mean", "std.error", "perc.025", "perc.975", "t stat."])
     summary.loc[:, "mean"] = data.mean(axis=0)
     summary.loc[:, "std.error"] = data.std(axis=0)
     summary.loc[:, "perc.025"] = data.quantile(0.025, axis=0)
@@ -43,11 +43,11 @@ class BootstrapProcess(Process):
         self.__iterations = iterations
 
     def run(self):
-        weights = pd.DataFrame(columns=self.__data.columns)
-        r_squared = pd.DataFrame(columns=self.__inner_model.r_squared().index)
-        total_effects = pd.DataFrame(columns=self.__inner_model.effects().index)
-        paths = pd.DataFrame(columns=self.__inner_model.effects().index)
-        loadings = pd.DataFrame(columns=self.__data.columns)
+        weights = pd.DataFrame(columns=self.__data.columns, dtype="float")
+        r_squared = pd.DataFrame(columns=self.__inner_model.r_squared().index, dtype="float")
+        total_effects = pd.DataFrame(columns=self.__inner_model.effects().index, dtype="float")
+        paths = pd.DataFrame(columns=self.__inner_model.effects().index, dtype="float")
+        loadings = pd.DataFrame(columns=self.__data.columns, dtype="float")
 
         observations = self.__data.shape[0]
         estimator = Estimator(self.__config)
@@ -58,8 +58,10 @@ class BootstrapProcess(Process):
                 weights = pd.concat([weights, _weights.T], ignore_index = True)
                 inner_model = im.InnerModel(self.__config.path(), _scores)
                 r_squared = pd.concat([r_squared, inner_model.r_squared().to_frame().T], ignore_index=True)
-                total_effects = pd.concat([total_effects, inner_model.effects().loc[:, "total"].to_frame().T], ignore_index=True)
-                paths = pd.concat([paths, inner_model.effects().loc[:, "direct"].to_frame().T], ignore_index=True)
+                total_effects = pd.concat([total_effects,
+                                           inner_model.effects().loc[:, "total"].to_frame().T], ignore_index=True)
+                paths = pd.concat([paths,
+                                   inner_model.effects().loc[:, "direct"].to_frame().T], ignore_index=True)
                 loadings = pd.concat([loadings,
                                       (_scores.apply(lambda s: _final_data.corrwith(s)) * self.__config.odm(self.__config.path())).sum(axis=1).to_frame().T], ignore_index=True)
             except:
@@ -80,11 +82,11 @@ class Bootstrap:
     """
     def __init__(self, config: c.Config, data: pd.DataFrame, inner_model: im.InnerModel, outer_model: om.OuterModel,
                  calculator: WeightsCalculatorFactory, iterations: int, num_processes: int):
-        weights = pd.DataFrame(columns=data.columns)
-        r_squared = pd.DataFrame(columns=inner_model.r_squared().index)
-        total_effects = pd.DataFrame(columns=inner_model.effects().index)
-        paths = pd.DataFrame(columns=inner_model.effects().index)
-        loadings = pd.DataFrame(columns=data.columns)
+        weights = pd.DataFrame(columns=data.columns, dtype="float")
+        r_squared = pd.DataFrame(columns=inner_model.r_squared().index, dtype="float")
+        total_effects = pd.DataFrame(columns=inner_model.effects().index, dtype="float")
+        paths = pd.DataFrame(columns=inner_model.effects().index, dtype="float")
+        loadings = pd.DataFrame(columns=data.columns, dtype="float")
 
         queue = Queue()
         processes = []
